Rename ArticlesListSection folder to ArticlesSection and Create DocumentsSection folder
