+++ next/src/components/layouts/Sections.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 
 import AccordionSection from '@/src/components/sections/AccordionSection'
-import ArticlesSection from '@/src/components/sections/ArticlesListSection/ArticlesSection'
+import ArticlesSection from '@/src/components/sections/ArticlesSection/ArticlesSection'
 import BannerSection from '@/src/components/sections/BannerSection'
 import CalculatorSection from '@/src/components/sections/CalculatorSection_Deprecated/CalculatorSection_Deprecated'
 import ColumnedTextSection from '@/src/components/sections/ColumnedTextSection'
@@ -9,7 +9,7 @@ import ColumnsSection from '@/src/components/sections/ColumnsSection'
 import ComparisonSection from '@/src/components/sections/ComparisonSection'
 import ContactsSection from '@/src/components/sections/ContactsSection'
 import DividerSection from '@/src/components/sections/DividerSection'
-import DocumentsSection from '@/src/components/sections/DocumentsSection'
+import DocumentsSection from '@/src/components/sections/DocumentsSection/DocumentsSection'
 import FaqCategoriesSection from '@/src/components/sections/FaqCategoriesSection'
 import FaqsSection from '@/src/components/sections/FaqsSection'
 import FileListSection from '@/src/components/sections/FileListSection'
