+++ next/src/components/sections/DocumentsSection/DocumentsSection.tsx
@@ -5,7 +5,7 @@ import React, { Fragment } from 'react'
 import DocumentRowCard from '@/src/components/cards/DocumentRowCard'
 import HorizontalDivider from '@/src/components/common/Divider/HorizontalDivider'
 import SectionContainer from '@/src/components/layouts/SectionContainer'
-import DocumentsAll from '@/src/components/sections/DocumentsAll'
+import DocumentsAll from '@/src/components/sections/DocumentsSection/DocumentsAll'
 import { DocumentsSectionFragment } from '@/src/services/graphql'
 import { formatDate } from '@/src/utils/formatDate'
 import { formatFileExtension } from '@/src/utils/formatFileExtension'
